+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.1.4] - 2024-08-12
+### Add
+* Add component to check ads available
+* Change signature of methods: StorageGet, StorageGetKeys StorageSet. Now storage should be used only via Storage class
+
 ## [2.1.3] - 2024-08-11
 ### Bugfix/Refactoring
 * Fixed issue with multiple requests of a single type are not handled properly